[readline] carry of #181

+++ readline/plan.sh
@@ -1,12 +1,15 @@
 pkg_name=readline
 pkg_origin=core
-_base_version=6.3
-pkg_version=${_base_version}.8
-pkg_license=('gplv3+')
-_url_base=http://ftp.gnu.org/gnu/$pkg_name
-pkg_source=$_url_base/${pkg_name}-${_base_version}.tar.gz
-pkg_dirname=${pkg_name}-$_base_version
-pkg_shasum=56ba6071b9462f980c5a72ab0023893b65ba6debb4eeb475d7a563dc65cafd43
+_base_version=7.0
+pkg_version=${_base_version}.0
+pkg_maintainer="The Habitat Maintainers <humans@habitat.sh>"
+pkg_license=('GPL-3.0')
+pkg_description="The GNU Readline library provides a set of functions for use by applications that allow users to edit command lines as they are typed in."
+_url_base="http://ftp.gnu.org/gnu/${pkg_name}"
+pkg_upstream_url="https://www.gnu.org/"
+pkg_source="${_url_base}/${pkg_name}-${_base_version}.tar.gz"
+pkg_dirname="${pkg_name}-${_base_version}"
+pkg_shasum=750d437185286f40a369e1e4f4764eda932b9459b5ec9a731628393dd3d32334
 pkg_deps=(core/glibc core/ncurses)
 pkg_build_deps=(core/coreutils core/diffutils core/patch core/make core/gcc core/bison core/grep)
 pkg_include_dirs=(include)
@@ -21,7 +24,7 @@ do_begin() {
 
   # Source a file containing an array of patch URLs and an array of patch file
   # shasums
-  source $PLAN_CONTEXT/readline-patches.sh
+  source "${PLAN_CONTEXT}/readline-patches.sh"
 }
 
 do_download() {
@@ -31,7 +34,7 @@ do_download() {
   # skip re-downloading if already present and verified
   for i in $(seq 0 $((${#_patch_files[@]} - 1))); do
     p="${_patch_files[$i]}"
-    download_file $p $(basename $p) ${_patch_shasums[$i]}
+    download_file "$p" "$(basename "$p")" "${_patch_shasums[$i]}"
   done; unset i p
 }
 
@@ -40,7 +43,7 @@ do_verify() {
 
   # Verify all patch files against their shasums
   for i in $(seq 0 $((${#_patch_files[@]} - 1))); do
-    verify_file $(basename ${_patch_files[$i]}) ${_patch_shasums[$i]}
+    verify_file "$(basename "${_patch_files[$i]}")" "${_patch_shasums[$i]}"
   done; unset i
 }
 
@@ -49,8 +52,8 @@ do_prepare() {
 
   # Apply all patch files to the extracted source
   for p in "${_patch_files[@]}"; do
-    build_line "Applying patch $(basename $p)"
-    patch -p0 -i $HAB_CACHE_SRC_PATH/$(basename $p)
+    build_line "Applying patch $(basename "$p")"
+    patch -p0 -i "${HAB_CACHE_SRC_PATH}/$(basename "$p")"
   done
 
   # This patch is to make sure that `libncurses' is among the `NEEDED'
@@ -62,14 +65,14 @@ do_prepare() {
   # Thanks to:
   # https://github.com/NixOS/nixpkgs/blob/release-15.09/pkgs/development/libraries/readline/link-against-ncurses.patch
   build_line "Applying patch link-against-ncurses.patch"
-  patch -p1 -i $PLAN_CONTEXT/link-against-ncurses.patch
+  patch -p1 -i "${PLAN_CONTEXT}/../readline/link-against-ncurses.patch"
 }
 
 do_install() {
   do_default_install
 
   # An empty `bin/` directory gets made, which we don't need and is confusing
-  rm -rf $pkg_prefix/bin
+  rm -rf "${pkg_prefix:?}/bin"
 }
